Fix get_random_file_offset() and get_byte_range_from_node()
Fix get_random_file_offset() by updating the code so that
other_thread_data is in this format:
{'172.16.1.10': ('n1.f1.txt', (0, 200))}

Fix get_byte_range_from_node() so that the condition where
all_offsets = [(0, 50)] and curr_offset = (0, 50) will return (0, 50).
Previously, it was returning None.

TODO: create a file's znode under /node-files if the file is not in
current_files.